+++ sktime/forecasting/compose/_pipeline.py
@@ -7,11 +7,11 @@ __all__ = ["TransformedTargetForecaster"]
 from itertools import islice
 from sklearn.base import clone
 from sktime.base import BaseComposition
-from sktime.forecasting.base.base import BaseForecaster
-from sktime.forecasting.base.sktime import BaseSktimeForecaster
-from sktime.forecasting.base.meta import MetaForecasterMixin
-from sktime.forecasting.base.sktime import OptionalForecastingHorizonMixin
-from sktime.forecasting.base.base import DEFAULT_ALPHA
+from sktime.forecasting.base._base import BaseForecaster
+from sktime.forecasting.base._sktime import BaseSktimeForecaster
+from sktime.forecasting.base._meta import MetaForecasterMixin
+from sktime.forecasting.base._sktime import OptionalForecastingHorizonMixin
+from sktime.forecasting.base._base import DEFAULT_ALPHA
 from sktime.transformers.detrend._base import BaseSeriesToSeriesTransformer
 from sktime.utils.validation.forecasting import check_y
