+++ sktime/forecasting/model_selection/_tune.py
@@ -21,8 +21,8 @@ from sklearn.model_selection import check_cv, ParameterGrid
 from sklearn.model_selection._search import _check_param_grid
 from sklearn.model_selection._validation import _aggregate_score_dicts
 from sklearn.utils.metaestimators import if_delegate_has_method
-from sktime.forecasting.base.base import BaseForecaster
-from sktime.forecasting.base.base import DEFAULT_ALPHA
+from sktime.forecasting.base._base import BaseForecaster
+from sktime.forecasting.base._base import DEFAULT_ALPHA
 from sktime.utils.exceptions import NotFittedError
 from sktime.utils.validation.forecasting import check_cv
 from sktime.utils.validation.forecasting import check_scoring
